move around SPI lib files
